Refactor with typecheck imports
