+++ .github/workflows/test_benchmarks.yaml
@@ -26,3 +26,22 @@ jobs:
           python -m pip install ".[tests]"
       - name: Run unit tests
         run: pytest
+
+  check_code_quality:
+    name: Check code quality
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Setup Python environment
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.9
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install ".[quality]"
+      - name: Code quality
+        run: |
+          make quality
+          make typecheck
