+++ benchmarks/raft/evaluation.py
@@ -1,26 +1,6 @@
-from typing import List, Optional, TypedDict, Union
-
 from datasets import load_dataset, load_metric
 
-
-class Metric(TypedDict):
-    name: str
-    type: str
-    value: Union[float, Optional[dict]]
-
-
-class Task(TypedDict):
-    name: str
-    type: str
-    metrics: List[Metric]
-
-
-class Result(TypedDict):
-    task: Task
-
-
-class Evaluation(TypedDict):
-    results: List[Result]
+from evaluate import Evaluation, Metric, Result, Task
 
 
 def convert_labels_to_ids(example):
