--- src/evaluate/__init__.py
@@ -1,4 +1,0 @@
-# flake8: noqa
-# coding=utf-8
-
-from .schemas import Evaluation, Metric, Task
